refactor: moved secrets_fetcher.py to /autograder/core/utils
